+++ inst/doc/introduction-to-ggfittext.R
@@ -1,37 +1,37 @@
-## ---- fig.width = 4, fig.height = 4--------------------------------------
+## ---- fig.width = 4, fig.height = 4-------------------------------------------
 library(ggfittext)
 library(ggplot2)
 ggplot(animals, aes(x = type, y = flies, label = animal)) +
   geom_tile(fill = "white", colour = "black") +
   geom_fit_text()
 
-## ---- fig.width = 4, fig.height = 4--------------------------------------
+## ---- fig.width = 4, fig.height = 4-------------------------------------------
 ggplot(animals, aes(x = type, y = flies, label = animal)) +
   geom_tile(fill = "white", colour = "black") +
   geom_fit_text(reflow = TRUE)
 
-## ---- fig.width = 4, fig.height = 4--------------------------------------
+## ---- fig.width = 4, fig.height = 4-------------------------------------------
 ggplot(animals, aes(x = type, y = flies, fill = mass, label = animal)) +
   geom_tile(fill = "white", colour = "black") +
   geom_fit_text(reflow = TRUE, grow = TRUE)
 
-## ---- fig.width = 4, fig.height = 4--------------------------------------
+## ---- fig.width = 4, fig.height = 4-------------------------------------------
 ggplot(animals, aes(x = type, y = flies, label = animal)) +
   geom_tile(fill = "white", colour = "black") +
   geom_fit_text(place = "topleft", reflow = TRUE)
 
-## ---- fig.width = 4, fig.height = 4--------------------------------------
+## ---- fig.width = 4, fig.height = 4-------------------------------------------
 ggplot(altitudes, aes(x = craft, y = altitude, label = altitude)) +
   geom_col() +
   geom_bar_text()
 
-## ---- fig.width = 4, fig.height = 4--------------------------------------
+## ---- fig.width = 4, fig.height = 4-------------------------------------------
 ggplot(coffees, aes(x = coffee, y = proportion, label = ingredient,
                     fill = ingredient)) +
   geom_col(position = "stack") +
   geom_bar_text(position = "stack", grow = TRUE, reflow = TRUE)
 
-## ---- fig.width = 4, fig.height = 4--------------------------------------
+## ---- fig.width = 4, fig.height = 4-------------------------------------------
 ggplot(coffees, aes(x = coffee, y = proportion, label = ingredient,
                     fill = ingredient)) +
   geom_col(position = "dodge") +
@@ -39,13 +39,35 @@ ggplot(coffees, aes(x = coffee, y = proportion, label = ingredient,
                 place = "left") +
   coord_flip()
 
-## ---- fig.width = 4, fig.height = 4--------------------------------------
+## ---- fig.width = 4, fig.height = 4-------------------------------------------
 ggplot(presidential, aes(ymin = start, ymax = end, x = party, label = name)) +
   geom_fit_text(grow = TRUE) +
   geom_errorbar(alpha = 0.5)
 
-## ---- fig.width = 4, fig.height = 4--------------------------------------
+## ---- fig.width = 4, fig.height = 4-------------------------------------------
+ggplot(gold, aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax, 
+                 label = label, fill = line)) +
+  geom_rect() +
+  coord_polar() +
+  geom_fit_text(min.size = 0, grow = TRUE) +
+  scale_fill_gradient(low = "#fee391", high = "#238443")
+
+## ---- fig.width = 4, fig.height = 4-------------------------------------------
 ggplot(animals, aes(x = type, y = flies, fill = mass, label = animal)) +
   geom_tile() +
   geom_fit_text(reflow = TRUE, grow = TRUE, contrast = TRUE)
+
+## ----echo = FALSE, fig.width = 4, fig.height = 0.5----------------------------
+ggplot(data.frame(label = "fullheight = TRUE", xmin = 0.25, xmax = 0.75, ymin = 0.25, ymax = 0.75), 
+       aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax, label = label)) +
+  geom_rect() +
+  geom_fit_text(grow = TRUE, fullheight = TRUE, family = "mono") +
+  theme(axis.text = element_blank(), axis.ticks = element_blank(), panel.background = element_blank(), plot.title = element_text(hjust = 0.5))
+
+## ----echo = FALSE, fig.width = 4, fig.height = 0.5----------------------------
+ggplot(data.frame(label = "fullheight = FALSE", xmin = 0.25, xmax = 0.75, ymin = 0.25, ymax = 0.75), 
+       aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax, label = label)) +
+  geom_rect() +
+  geom_fit_text(grow = TRUE, fullheight = FALSE, family = "mono") +
+  theme(axis.text = element_blank(), axis.ticks = element_blank(), panel.background = element_blank(), plot.title = element_text(hjust = 0.5))