+++ tests/testthat/test_plots.R
@@ -1,5 +1,5 @@
 library(ggplot2)
-library(ggfittext)
+grid::current.viewport()
 
 yeats <- data.frame(
   xmin = c(0, 4, 6, 4, 4, 5, 5.5, 5,   5,    5.25, 5.25),
@@ -10,6 +10,93 @@ yeats <- data.frame(
             "falcon", "cannot", "hear"),
   angle = c(0, 315, 270, 225, 180, 135, 90, 45, 0, 315, 270)
 )
+
+context("geom_bar_text()")
+
+test_that("plots look the way they should", {
+
+  vdiffr::expect_doppelganger("coord_flip", {
+    ggplot(altitudes, aes(x = craft, y = altitude, label = altitude)) +
+      coord_flip() +
+      geom_col() +
+      geom_bar_text()
+  })
+
+  vdiffr::expect_doppelganger("implied flip", {
+    ggplot(altitudes, aes(y = craft, x = altitude, label = altitude)) +
+      geom_col() +
+      geom_bar_text()
+  })
+
+  vdiffr::expect_doppelganger("coord_flip with grow", {
+    ggplot(altitudes, aes(x = craft, y = altitude, label = altitude)) +
+      coord_flip() +
+      geom_col() +
+      geom_bar_text(grow = TRUE)
+  })
+
+  vdiffr::expect_doppelganger("implied flip with grow", {
+    ggplot(altitudes, aes(y = craft, x = altitude, label = altitude)) +
+      geom_col() +
+      geom_bar_text(grow = TRUE)
+  })
+
+  vdiffr::expect_doppelganger("coord_flip with place middle", {
+    ggplot(altitudes, aes(x = craft, y = altitude, label = altitude)) +
+      coord_flip() +
+      geom_col() +
+      geom_bar_text(place = "middle")
+  })
+
+  vdiffr::expect_doppelganger("implied flip with place middle", {
+    ggplot(altitudes, aes(y = craft, x = altitude, label = altitude)) +
+      geom_col() +
+      geom_bar_text(place = "middle")
+  })
+
+  vdiffr::expect_doppelganger("coord_flip with dodge", {
+    ggplot(coffees, aes(x = coffee, y = proportion, label = ingredient,
+                        fill = ingredient)) +
+      geom_col(position = "dodge") +
+      geom_bar_text(position = "dodge") +
+      coord_flip()
+  })
+
+  vdiffr::expect_doppelganger("coord_flip with dodge and grow and reflow", {
+    ggplot(coffees, aes(x = coffee, y = proportion, label = ingredient,
+                        fill = ingredient)) +
+      geom_col(position = "dodge") +
+      geom_bar_text(position = "dodge", grow = TRUE, reflow = TRUE, 
+                    place = "left") +
+      coord_flip()
+  })
+
+  vdiffr::expect_doppelganger("implied flip with dodge", {
+    ggplot(coffees, aes(y = coffee, x = proportion, label = ingredient,
+                        fill = ingredient)) +
+      geom_col(position = "dodge") +
+      geom_bar_text(position = "dodge")
+  })
+
+  vdiffr::expect_doppelganger("implied flip with dodge and grow and reflow", {
+    ggplot(coffees, aes(y = coffee, x = proportion, label = ingredient,
+                        fill = ingredient)) +
+      geom_col(position = "dodge") +
+      geom_bar_text(position = "dodge", grow = TRUE, reflow = TRUE, 
+                    place = "left")
+  })
+
+
+  vdiffr::expect_doppelganger("implied doesn't misfire with integer y", {
+    c2 <- coffees
+    c2$proportion <- 1:6
+    ggplot(c2, aes(x = coffee, y = proportion, label = ingredient,
+                        fill = ingredient)) +
+    geom_col(position = "dodge") +
+    geom_bar_text(position = "dodge")
+  })
+
+})
 
 context("visual tests of plots")
 
@@ -35,7 +122,7 @@ test_that("plots look the way they should", {
   })
 
   vdiffr::expect_doppelganger("Reflowing and growing", {
-    ggplot(animals, aes(x = type, y = flies, fill = mass, label = animal)) +
+    ggplot(animals, aes(x = type, y = flies, label = animal)) +
       geom_tile(fill = "white", colour = "black") +
       geom_fit_text(reflow = TRUE, grow = TRUE, size = 48)
   })
@@ -81,8 +168,27 @@ test_that("plots look the way they should", {
   })
 
   vdiffr::expect_doppelganger("Blank labels", {
-    presidential$name[1] <- ""
-    ggplot(presidential, aes(ymin = start, ymax = end, label = name, x = party)) +
+    pressies <- presidential
+    pressies$name[1] <- ""
+    ggplot(pressies, aes(ymin = start, ymax = end, label = name, x = party)) +
         geom_fit_text(grow = TRUE)
   })
+
+  vdiffr::expect_doppelganger("Contrast against default bar colour", {
+    ggplot(altitudes, aes(x = craft, y = altitude, label = altitude)) +
+      geom_col() +
+      geom_bar_text(contrast = TRUE)
+  })
+
+  vdiffr::expect_doppelganger("Contrasting works with non-black text", {
+    ggplot(animals, aes(x = type, y = flies, fill = mass, label = animal)) +
+      geom_tile() +
+      geom_fit_text(colour = "thistle", reflow = TRUE, grow = TRUE, 
+                    contrast = TRUE) +
+       scale_fill_gradientn(
+         colours = c("red","yellow","green","lightblue","darkblue"),
+         values = c(1.0,0.8,0.6,0.4,0.2,0)
+       ) 
+  })
+
 })